Local embedders implementation
